// client/src/types.tsx
export type UserProps = {
    name: string;
    username: string;
    avatar: string;
    online: boolean;
  };
  
  export type MessageProps = {
    id: string;
    content: string;
    timestamp: string;
    unread?: boolean;
    sender: UserProps | 'You';
    // attachment?: {
    //   fileName: string;
    //   type: string;
    //   size: string;
    // };
  };
  
  export type ChatProps = {
    id: string;
    sender: UserProps;
    messages: MessageProps[];
  };

  export type myChatProps = {
    conversation: string; //(id)
    from: string; //(sender)
    message: string; // messages:
    timestamp: Date;
    to:string;
  }